Refactor popup initialization
The popup initialization code was duplicated between `notification.js`
and `ui.js`. It has been moved into `popup-core` instead. Some changes
were made to `popup-core` to improve readability as well.
